:hammer: refactor: Seperate dtos to request and response directory
