+++ backend/src/blocked/blocked.controller.ts
@@ -1,3 +1,9 @@
+import { ErrorResponseDto } from '../common/dto/error-response.dto';
+import { SuccessResponseDto } from '../common/dto/success-response.dto';
+
+import { BlockedService } from './blocked.service';
+import { BlockedUserResponseDto } from './dto/response/blocked-user-response.dto';
+
 import { Controller, Delete, Get, Headers, HttpCode, HttpStatus, Param, Post, Query } from '@nestjs/common';
 import {
   ApiConflictResponse,
@@ -9,12 +15,6 @@ import {
   ApiQuery,
   ApiTags,
 } from '@nestjs/swagger';
-
-import { ErrorResponseDto } from '../common/dto/error-response.dto';
-import { SuccessResponseDto } from '../common/dto/success-response.dto';
-
-import { BlockedService } from './blocked.service';
-import { BlockedUserResponseDto } from './dto/blocked-user-response.dto';
 
 @ApiTags('blocked')
 @Controller('blocked')
