+++ backend/src/blocked/blocked.service.ts
@@ -1,3 +1,11 @@
+import { BLOCKED_USER_LIMIT } from '../common/constant';
+import { SuccessResponseDto } from '../common/dto/success-response.dto';
+import { BlockedUser } from '../entity/blocked-user.entity';
+import { Friendship } from '../entity/friendship.entity';
+import { UserService } from '../user/user.service';
+
+import { BlockedUserResponseDto } from './dto/response/blocked-user-response.dto';
+
 import {
   BadRequestException,
   ConflictException,
@@ -7,14 +15,6 @@ import {
 } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-
-import { BLOCKED_USER_LIMIT } from '../common/constant';
-import { SuccessResponseDto } from '../common/dto/success-response.dto';
-import { BlockedUser } from '../entity/blocked-user.entity';
-import { Friendship } from '../entity/friendship.entity';
-import { UserService } from '../user/user.service';
-
-import { BlockedUserResponseDto } from './dto/blocked-user-response.dto';
 
 @Injectable()
 export class BlockedService {
