+++ backend/src/blocked/dto/response/blocked-user-response.dto.ts
@@ -1,6 +1,6 @@
 import { BlockedUserResponse } from '@/types/blocked/response';
 
-import { UserInfoDto } from '../../user/dto/user-info.dto';
+import { UserInfoDto } from '../../../user/dto/user-info.dto';
 export class BlockedUserResponseDto implements BlockedUserResponse {
   /**
    * 차단한 유저 리스트
