+++ backend/src/friend/dto/response/friend-response.dto.ts
@@ -1,6 +1,6 @@
 import { FriendResponse } from '@/types/friend/response';
 
-import { UserInfoDto } from '../../user/dto/user-info.dto';
+import { UserInfoDto } from '../../../user/dto/user-info.dto';
 
 class FriendInformation {
   /**
