+++ backend/src/friend/dto/response/requested-friend-response.dto.ts
@@ -1,6 +1,6 @@
 import { RequestedFriendsResponse } from '@/types/friend/response';
 
-import { UserInfoDto } from '../../user/dto/user-info.dto';
+import { UserInfoDto } from '../../../user/dto/user-info.dto';
 
 export class RequestedFriendsResponseDto implements RequestedFriendsResponse {
   /**
