+++ backend/src/user/dto/request/user-image-request.dto.ts
