+++ backend/src/user/dto/response/user-nickname-response.dto.ts
